Finished whitelist add command. Roles added to member when opening specific games.
Ignore the list of servers record.
Command created so admins can add a whitelisted app with either a name or a role.
This is done by using !gmadd [Game Name] [@Role OR Desired name of role].
If the bot cannot find a role by the name the admin has entered it will create the role.
Moved creation of whitelisted app into a function to make command section cleaner.