Records time user has a presence for.
Records the time the user has a program open for, before they change to another, done by taking the time the user opened the program and when their presence changes to a different program, tiem is then reducted and added to a stats file for the server they are in.

Does not appear to work properly when a user is in multiple servers yet.