Starting work on pages systems for stats
Allows user to view different pages of their stats as all stats do not fit in one graph (or one page).
Use case of the !gmstats command will still be the same but an optional number can be put on the end to specify a page, if no number is specified then the first page is shown.
Added file to gitignore.