Added removing of day, week and month content
Minor changes to getting stats file paths as these take ID's now and not the guild object.

Day works by checking if the curernt time is 00:00.
Week works by checking if it is 00:00 on a Monday.
Month works by checking if it is 00:00 on the 1st.
All these checks are checked every minute.